Further refactor getValidCoordinates for create, update
