+++ server/app/routes/api/v1/va/ambassadors.js
@@ -2,13 +2,13 @@ import { Router } from 'express';
 import format from 'string-format';
 import { v4 as uuidv4 } from 'uuid';
 
-import { normalizePhone } from '../../../../lib/normalizers';
+import { getValidCoordinates, normalizePhone } from '../../../../lib/normalizers';
 import { ValidationError } from '../../../../lib/errors';
 import ambassadorsSvc from '../../../../services/ambassadors';
 import { error } from '../../../../services/errors';
 
 import {
-  _204, _400, _401, _403, _404, geoCode
+  _204, _400, _401, _403, _404
 } from '../../../../lib/utils';
 
 import {
@@ -35,15 +35,12 @@ async function createAmbassador(req, res) {
       return error(400, res, "Invalid payload, ambassador cannot be created");
     }
 
+    let coordinates, address;
     try {
       await assertUserPhoneAndEmail('Ambassador', req.body.phone, req.body.email);
+      [coordinates, address] = await getValidCoordinates(req.body.address);
     } catch (err) {
       return error(400, res, err.message, req.body);
-    }
-
-    let coordinates = await geoCode(req.body.address);
-    if (!coordinates) {
-      return error(400, res, "Our system doesn't recognize that address. Please try again.");
     }
 
     new_ambassador = await req.neode.create('Ambassador', {
@@ -52,7 +49,7 @@ async function createAmbassador(req, res) {
       last_name: req.body.last_name || null,
       phone: normalizePhone(req.body.phone),
       email: req.body.email || null,
-      address: JSON.stringify(req.body.address, null, 2),
+      address: JSON.stringify(address, null, 2),
       quiz_results: JSON.stringify(req.body.quiz_results, null, 2) || null,
       approved: false,
       locked: false,
